🔍 refactor. badge element 정리
