+++ src/elements/badge/badge.type.ts
@@ -1,0 +1,4 @@
+export interface BadgeType {
+  children?: React.ReactNode;
+  type: "new" | "best" | "sale" | "soldout" | "ad";
+}
